@@ -27,75 +29,93 @@
     "node": ">= 6.2.1"
   },
   "dependencies": {
-    "backbone-session": "^1.0.1",
+    "app-module-path": "^2.2.0",
     "bcrypt-nodejs": "0.0.3",
-    "bluebird": "^3.4.6",
-    "body-parser": "~1.15.2",
-    "config": "^1.21.0",
+    "bluebird": "^3.5.0",
+    "body-parser": "~1.17.1",
+    "config": "^1.26.1",
+    "cookie": "^0.3.1",
     "cookie-parser": "^1.4.3",
-    "debug": "~2.2.0",
-    "errorhandler": "^1.4.3",
-    "express": "~4.14.0",
+    "debug": "~2.6.6",
+    "errorhandler": "^1.5.0",
+    "express": "~4.15.2",
     "express-promise-router": "^1.1.1",
-    "i18next": "^3.4.3",
-    "i18next-express-middleware": "^1.0.2",
-    "i18next-node-fs-backend": "^0.1.2",
-    "jsonwebtoken": "^7.1.9",
-    "method-override": "^2.3.6",
-    "moment": "^2.15.2",
-    "mongodb": "~2.2.10",
-    "morgan-debug": "^1.0.0",
+    "graphql": "^0.9.6",
+    "graphql-server-express": "^0.7.2",
+    "graphql-subscriptions": "^0.3.1",
+    "graphql-tools": "^0.11.0",
+    "i18next": "^8.1.0",
+    "i18next-express-middleware": "^1.0.5",
+    "i18next-node-fs-backend": "^1.0.0",
+    "jsonwebtoken": "^7.4.0",
+    "lodash": "^4.17.4",
+    "method-override": "^2.3.8",
+    "moment": "^2.18.1",
+    "mongodb": "~2.2.26",
+    "morgan": "^1.8.1",
+    "morgan-debug": "^2.0.0",
     "passport": "^0.3.2",
-    "passport-jwt": "^2.1.0",
+    "passport-anonymous": "^1.0.1",
+    "passport-jwt": "^2.2.1",
     "permission": "^1.0.4",
-    "serve-favicon": "~2.3.0",
+    "serve-favicon": "~2.4.2",
     "stream-to-promise": "^2.2.0",
-    "underscore": "1.8.3"
+    "subscriptions-transport-ws": "^0.6.0"
   },
   "devDependencies": {
-    "apidoc": "^0.16.1",
-    "autoprefixer": "^6.5.0",
-    "babel-core": "^6.16.0",
-    "babel-eslint": "^7.0.0",
-    "babel-loader": "6.2.5",
+    "apollo-client": "^1.2.2",
+    "apidoc": "^0.17.5",
+    "autoprefixer": "^7.0.1",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "7.0.0",
+    "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.16.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-es2015-loose": "^8.0.0",
-    "babel-preset-stage-1": "^6.16.0",
-    "babel-runtime": "^6.11.6",
-    "backbone": "^1.3.3",
-    "backbone-deep-model": "^0.11.0",
-    "backbone-validation": "^0.11.5",
-    "backbone.marionette": "^2.4.7",
-    "backbone.stickit": "^0.9.2",
+    "babel-preset-stage-1": "^6.24.1",
+    "babel-runtime": "^6.23.0",
+    "clipboard-copy": "^1.2.0",
     "connect-flash": "^0.1.1",
-    "copy-webpack-plugin": "^3.0.1",
+    "copy-webpack-plugin": "^4.0.1",
     "crypto-browserify": "^3.11.0",
-    "css-loader": "^0.25.0",
+    "css-loader": "^0.28.1",
+    "download-as-file": "^1.1.0",
     "ejs-loader": "^0.3.0",
-    "eslint": "^3.6.1",
+    "eslint": "^3.19.0",
     "eslint-plugin-nodeca": "^1.0.3",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "hammerjs": "^2.0.8",
-    "html-loader": "^0.4.4",
-    "html-webpack-plugin": "^2.22.0",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.11.1",
+    "html-loader": "^0.4.5",
+    "html-webpack-plugin": "^2.28.0",
     "i18next-resource-store-loader": "0.1.1",
-    "img-loader": "^1.3.1",
-    "jade": "^1.11.0",
-    "jade-loader": "^0.8.0",
-    "jquery": "^3.1.1",
+    "img-loader": "^2.0.0",
+    "jquery": "^3.2.1",
     "json-loader": "^0.5.4",
-    "materialize-css": "^0.97.7",
-    "node-sass": "^3.10.0",
-    "postcss-loader": "^0.13.0",
-    "rimraf": "^2.5.4",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.16.1",
+    "json2csv": "^3.7.3",
+    "lodash-webpack-plugin": "^0.11.3",
+    "node-sass": "^4.5.2",
+    "postcss-loader": "^2.0.1",
+    "pug": "^2.0.0-rc.1",
+    "pug-loader": "^2.3.0",
+    "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.3",
+    "style-loader": "^0.17.0",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^3.0.1",
+    "url-loader": "^0.5.8",
+    "vee-validate": "^2.0.0-rc.3",
+    "vue": "^2.3.3",
+    "vue-apollo": "^2.1.0-beta.8",
+    "vue-loader": "^12.0.4",
+    "vue-router": "^2.5.3",
+    "vue-template-compiler": "^2.3.3",
+    "vuetify": "^0.12.0",
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5",
     "yaml-loader": "^0.4.0"
   }
 }
