@@ -1,29 +1,11 @@
 'use strict';
 
-const debug = require('debug')('App:Models:Exception');
+import debug from 'debug';
+import i18n from 'i18next';
 
-class DuplicateKeyError extends Error {
-	constructor(err) {
-		super(err.message, err.id);
-		this.originalError = err;
-		this.name = 'DuplicateKeyError';
-		this.status = 400;
+const log = debug('App:Models:Exception');
 
-		const searchFieldRegexp = /E11000 duplicate key error collection:\s+\w+\.\w+\s+index:\s+(\w+)_\s+dup key/g;
-		const fields = searchFieldRegexp.exec(err.errmsg);
-		this.field = fields.length > 1 ? fields[1] : null;
-	}
-}
-
-class NotFoundError extends Error {
-	constructor(message) {
-		super(message);
-		this.name = 'NotFoundError';
-		this.status = 404;
-	}
-}
-
-class AuthorizationError extends Error {
+export class AuthorizationError extends Error {
 	constructor(message) {
 		super(message);
 		this.name = 'AuthorizationError';
@@ -31,16 +13,29 @@ class AuthorizationError extends Error {
 	}
 }
 
-module.exports.DuplicateKeyError = DuplicateKeyError;
-module.exports.NotFoundError = NotFoundError;
-module.exports.AuthorizationError = AuthorizationError;
-module.exports.processMongoException = processMongoException;
+export class DuplicateKeyError extends Error {
+	constructor(err, field) {
+		super(i18n.t('error:global.400.duplicate'), err.id);
+		this.originalError = err;
+		this.name = 'DuplicateKeyError';
+		this.status = 400;
+		this.field = field;
+	}
+}
 
-function processMongoException(err) {
+export class NotFoundError extends Error {
+	constructor(message) {
+		super(message);
+		this.name = 'NotFoundError';
+		this.status = 404;
+	}
+}
+
+export function processMongoException(err) {
 	if (err.name === 'MongoError' && err.code === 11000) {
-		throw new DuplicateKeyError(err);
+		throw new DuplicateKeyError(err, '_id');
 	}
 
-	debug(err);
+	log(err);
 	throw err;
 }
