@@ -1,12 +1,11 @@
 'use strict';
 
-const _ = require('underscore');
-const db = require('../utils/db');
-const i18n = require('i18next');
-const ObjectID = require('mongodb').ObjectID;
+import {isString, omit}  from 'lodash';
+import {promise as dbPromise} from 'server/utils/db';
+import i18n from 'i18next';
+import { ObjectID } from 'mongodb';
 
-const processMongoException = require('./exception').processMongoException;
-const NotFoundError = require('./exception').NotFoundError;
+import { processMongoException, NotFoundError } from './exception';
 
 /*
  * Model of a wallet line contains :
@@ -20,55 +19,51 @@ const NotFoundError = require('./exception').NotFoundError;
  *
  * And all other informations that a user want to store in the wallet line (depending on the wallet).
  */
-module.exports = function () {
-	return {
-		getLines(filter, offset, limit, sort) {
-			const find = {};
+export function getLines(filter, sort) {
+	const find = {};
 
-			if (_.isString(filter.user)) {
-				find.user = filter.user;
-			}
-
-			return db.promise.then(db => {
-				return Promise.fromCallback(cb => db.collection('walletlines').find(find, {_id: 1, type: 1, label: 1}).sort(sort).skip(offset).limit(limit).toArray(cb));
-			});
-		},
-
-		getLine(id, _rev) {
-			const query = {_id: new ObjectID(id)};
-			if (_rev !== undefined) {
-				query._rev = _rev;
-			}
-
-			return db.promise.then(db => {
-				return Promise.fromCallback(cb => db.collection('walletlines').findOne(query, cb));
-			}).then(_.partial(processNotFound, id));
-		},
-
-		saveLine(line) {
-			const cleanLine = _.omit(line, '_rev');
-			const revision = line._rev;
-			const query = {_id: new ObjectID(line._id)};
-			if (revision) {
-				query._rev = revision;
-			}
-
-			return db.promise.then(db => {
-				return Promise.fromCallback(cb => db.collection('walletlines').findOneAndUpdate(query, {$set: cleanLine, $inc: {_rev: 1}}, {returnOriginal: false, upsert: true}, cb)).then(doc => doc.value);
-			}).catch(processMongoException);
-		},
-
-		removeLine(id) {
-			return db.promise.then(db => {
-				return Promise.fromCallback(cb => db.collection('walletlines').deleteOne({_id: new ObjectID(id)}, cb));
-			}).then(_.partial(processNotFound, id));
-		}
-	};
-
-	function processNotFound(lineId, line) {
-		if (!line) {
-			throw new NotFoundError(i18n.t('error:line.404.lineNotFound', {lineId}));
-		}
-		return line;
+	if (isString(filter.user)) {
+		find.user = filter.user;
 	}
-};
+
+	return dbPromise.then(db => {
+		return Promise.fromCallback(cb => db.collection('walletlines').find(find).sort(sort).toArray(cb));
+	});
+}
+
+export function getLine(id, _rev) {
+	const query = { _id: new ObjectID(id) };
+	if (_rev !== undefined) {
+		query._rev = _rev;
+	}
+
+	return dbPromise.then(db => {
+		return Promise.fromCallback(cb => db.collection('walletlines').findOne(query, cb));
+	}).then(line => processNotFound(id, line));
+}
+
+export function saveLine(line) {
+	const cleanLine = omit(line, '_rev');
+	const revision = line._rev;
+	const query = { _id: new ObjectID(line._id) };
+	if (revision) {
+		query._rev = revision;
+	}
+
+	return dbPromise.then(db => {
+		return Promise.fromCallback(cb => db.collection('walletlines').findOneAndUpdate(query, { $set: cleanLine, $inc: { _rev: 1 } }, { returnOriginal: false, upsert: true }, cb)).then(doc => doc.value);
+	}).catch(processMongoException);
+}
+
+export function removeLine(id) {
+	return dbPromise.then(db => {
+		return Promise.fromCallback(cb => db.collection('walletlines').deleteOne({ _id: new ObjectID(id) }, cb));
+	}).then(line => processNotFound(id, line));
+}
+
+function processNotFound(lineId, line) {
+	if (!line) {
+		throw new NotFoundError(i18n.t('error:line.404.lineNotFound', { lineId }));
+	}
+	return line;
+}
