@@ -1,41 +1,38 @@
 'use strict';
 
-const debug = require('debug')('App:Service:LineTransaction');
-const NotFoundError = require('../models/exception').NotFoundError;
-const crypto = require('crypto');
+import debug from 'debug';
 
-module.exports = function (models, services) {
-	return {
-		getTransactions(user, params) {
-			debug(`Get all transaction of the user ${user._id}`);
-			if (params.limit === 0) {
-				return Promise.resolve([]);
-			}
+import crypto from 'crypto';
+import {getTransactions as getTransactionsModel, createTransaction as createTransactionModel} from 'server/models/transaction';
+import {transactionAdded} from 'server/services/subscriptions';
 
-			const filter = {};
-			filter.user = user._id;
-			filter.earliest = params.earliest;
+const log = debug('App:Service:LineTransaction');
 
-			return models.transaction.getTransactions(filter, params.offset, params.limit);
-		},
+export function getTransactions(user, params = {}) {
+	log(`Get all transaction of the user ${user._id} after ${params.earliest}`);
 
-		createTransaction(type, before, after) {
-			const base = after || before;
-			if (!base) {
-				return null;
-			}
+	const filter = {};
+	filter.user = user._id;
+	filter.earliest = params.earliest;
 
-			const transaction = {
-				type,
-				line: base._id,
-				user: base.user,
-				before: before,
-				after: after,
-				updatedAt: base.updatedAt,
-				sha512: after && crypto.createHash('sha512').update(after.encryption.informations.content).digest('hex')
-			};
+	return getTransactionsModel(filter);
+}
 
-			return models.transaction.create(transaction);
-		}
+export function createTransaction(type, before, after) {
+	const base = after || before;
+	if (!base) {
+		return null;
+	}
+
+	const transaction = {
+		type,
+		line: base._id,
+		user: base.user,
+		before: before,
+		after: after,
+		updatedAt: base.updatedAt,
+		sha512: after && crypto.createHash('sha512').update(after.encryption.informations.content).digest('hex')
 	};
-};
+
+	return createTransactionModel(transaction).tap(transactionAdded);
+}
