@@ -1,23 +1,35 @@
 'use strict';
 
-import '../css/passprotect.scss';
-
-import './bootstrap/materializecss';
+import {init as i18nInit} from 'i18next';
 import resBundle from 'i18next-resource-store-loader!nscommon/locales';
-import helpers from 'nscommon/helpers';
+import Promise from 'bluebird';
+import Vue from 'vue';
+import Vuetify from 'vuetify';
+import VueRouter from 'vue-router';
+import VueApollo from 'vue-apollo';
+import veeDictionary from 'vee-validate/dist/locale/fr';
+import VeeValidate, { Validator } from 'vee-validate';
 
-import application from 'nsclient/application';
-import router from 'nsclient/router';
-import {addRegions} from 'nsclient/common/region';
+import App from './App.vue';
+import Login from './user/Login.vue';
+import Register from './user/Register.vue';
+import Items from './items/Items.vue';
+import ItemModification from './items/ItemModification.vue';
+import ItemVisualisation from './items/ItemVisualisation.vue';
+import ItemCreation from './items/ItemCreation.vue';
+import About from './page/About.vue';
+import {apolloClient} from './utils/graphql';
+import {checkAuth} from './user/UserService';
 
-import 'nsclient/common/plugins';
-import 'nsclient/common/services/errorsService';
+Validator.addLocale(veeDictionary);
 
-import routesEventService from 'nsclient/common/services/routesEventService';
+Vue.use(Vuetify);
+Vue.use(VueRouter);
+Vue.use(VueApollo);
+Vue.use(VeeValidate, {errorBagName: 'veeErrors', locale: 'fr'});
 
-// Configure backbone
-Backbone.emulateHTTP = true;
-Backbone.$ = jQuery;
+// Set the promise as global
+window.Promise = Promise;
 
 const i18nOptions = {
 	resources: resBundle,
@@ -25,28 +37,34 @@ const i18nOptions = {
 	joinArrays: '+'
 };
 
-const i18nInitPromise = Promise.fromCallback((cb) => i18n.init(i18nOptions, cb)).then((t) => {
-	window.t = t;
-});
+const i18nInitPromise = Promise
+	.fromCallback((cb) => i18nInit(i18nOptions, cb))
+	.then(trans => {
+		window.trans = trans;
+		Vue.prototype.trans = trans;
+	});
 
-// Load Jade helper
-helpers(window);
+const routes = [
+	{path: '/about', component: About},
+	{path: '/items', component: Items},
+	{path: '/items/type/:type', component: ItemCreation, props: true},
+	{path: '/items/:id/edit', component: ItemModification, props: true},
+	{path: '/items/:id', component: ItemVisualisation, props: true},
+	{path: '/login', component: Login},
+	{path: '/register', component: Register},
+	{path: '*', redirect: '/items'}
+];
 
-// Set the application
-addRegions(application);
-
-application.on('start', () => {
-	router(application);
-
-	if (Backbone.history) {
-		Backbone.history.start();
-
-		if (application.getCurrentRoute() === '') {
-			routesEventService.trigger('items:list');
-		}
-	}
-});
+const router = new VueRouter({routes});
+const apolloProvider = new VueApollo({defaultClient: apolloClient});
 
 i18nInitPromise.then(() => {
-	application.start();
+	checkAuth();
+
+	new Vue({
+		el: '#app',
+		render: h => h(App),
+		router,
+		apolloProvider
+	});
 });
